backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
PyMySQL==1.1.2
bcrypt==4.3.0
python-decouple==3.8
Pillow==11.3.0
